Move color functions to their own package
