+++ pkg/color/color256rgb.go
@@ -1,4 +1,4 @@
-package pstree
+package color
 
 import "fmt"
