+++ pkg/color/color8.go
@@ -1,4 +1,4 @@
-package pstree
+package color
 
 import (
 	"fmt"
